Have RandGen inherit std::mt19937
This is kind of a hack because there is no point having that abstract
class inherit a concrete random generator. But it's clear we don't
need that home made random generator structure anymore so it's a step
toward getting rid of it.

The advantage is that the random generator returned by the singleton
randGen() can be now used in the STL distributions.