Disable persisting ccache in the ure build to avoid circleci error
The error in question is

Error applying workspace layer for job XXXX
Concurrent upstream jobs persisted the same file(s)

which seems to indicate that some shared ccache files get updated by
the URE and some other process running in parallel such as
Attention. It's strange that different projects, living in different
paths lead ccache to write on shared files. In ccache documentation
https://ccache.dev/manual/3.7.4.html Section Compiling in different
directories, we can read

"
This means that if you compile the same code in different locations,
you can’t share compilation results between the different build
directories since you get cache misses because of the absolute build
directory paths that are part of the hash.
"

which seems to contradicts this. Not sure if that's a bug on the part
of ccache, or maybe a bug on the part of circleci. But in the meantime
disabling persisting ccache for the URE fixes the problem, with the
inconvenient that compiling speed will degrade overtime.