Revert "Rework barrier to take a factory function."
This reverts commit 2ffbf08490066a60d11739d7c60101af4301e9bd.

This is spinning out of control. I have let Claude write this code,
and I've been naively commiting the fixes, because they smelled OK,
and I was trusting Claude to get it write. However, we are now on
day three of debugging crashes and hangs in this new code. First
step is to revert some of the increasingly baroque changes.